fix(fulfill-all-required-changes): all required changes are fulfilled
