+++ src/components/Universe/Graph/UI/NodeControls/index.tsx
@@ -164,7 +164,7 @@ export const NodeControls = memo(() => {
 
   const id = open ? 'simple-popover' : undefined
 
-  const isShowCreateTestButton = !!(selectedNode && selectedNode.node_type === 'function')
+  const isShowCreateTestButton = !!(selectedNode && selectedNode?.node_type?.toLowerCase() === 'function')
 
   return (
     <group ref={ref}>
@@ -199,8 +199,6 @@ export const NodeControls = memo(() => {
 
         {isShowCreateTestButton && (
           <CreateTestButton
-            backgroundColor={colors.createTestButton}
-            fontColor="black"
             left={2}
             onClick={() => {
               createBountyModal()
@@ -271,28 +269,6 @@ const IconButton = styled.div<ButtonProps>`
   box-shadow: 0px 2px 12px rgba(0, 0, 0, 0.5);
 `
 
-const CreateTestButton = styled.div<ButtonProps>`
-  position: fixed;
-  top: 40px;
-  left: ${(p: ButtonProps) => -53 + p.left}px;
-  width: 100px;
-  padding: 6px;
-  border-radius: 4px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: ${(p: ButtonProps) => (p.backgroundColor ? p.backgroundColor : '#000000bb')};
-  color: ${(p: ButtonProps) => (p.fontColor ? p.fontColor : '#ffffff')};
-  font-size: 14px;
-  font-family: Barlow;
-  font-weight: 600;
-  cursor: pointer;
-
-  &:hover {
-    transform: scale(1.05);
-  }
-`
-
 const PopoverOption = styled(Flex).attrs({
   direction: 'row',
   px: 12,
@@ -325,5 +301,26 @@ const PopoverWrapper = styled(Popover)`
     font-family: Barlow;
     font-size: 14px;
     font-weight: 500;
+  }
+`
+
+const CreateTestButton = styled.div<ButtonProps>`
+  position: fixed;
+  top: 40px;
+  left: ${(p: ButtonProps) => -53 + p.left}px;
+  width: 100px;
+  padding: 6px;
+  border-radius: 4px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: ${colors.createTestButton};
+  color: ${colors.black};
+  font-size: 14px;
+  font-family: Barlow;
+  font-weight: 600;
+  cursor: pointer;
+  &:hover {
+    transform: scale(1.05);
   }
 `
